requirements.txt
Flask
pytesseract
Pillow
transformers
torch
faiss-cpu
sentence-transformers
